\name{mclogit_fast-package}
\alias{mclogit_fast-package}
\alias{mclogit_fast}
\docType{package}
\title{
\packageTitle{mclogit_fast}
}
\description{
\packageDescription{mclogit_fast}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mclogit_fast}
\packageIndices{mclogit_fast}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mclogit_fast}

Maintainer: \packageMaintainer{mclogit_fast}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
